refactor ExtraField handling and remove file extension from import
