+++ __tests__/binary/index.ts
@@ -1,7 +1,7 @@
-export * from "./CentralDirectoryFileHeader.js";
-export * from "./DataDescriptor.js";
-export * from "./EndOfCentralDirectoryRecord.js";
-export * from "./ExtraField.js";
-export * from "./LocalFileHeader.js";
-export * from "./Zip64EndOfCentralDirectoryRecord.js";
-export * from "./Zip64EndOfCentralDirectoryLocator.js";
+export * from "./CentralDirectoryFileHeader";
+export * from "./DataDescriptor";
+export * from "./EndOfCentralDirectoryRecord";
+export * from "./ExtraField/index";
+export * from "./LocalFileHeader";
+export * from "./Zip64EndOfCentralDirectoryLocator";
+export * from "./Zip64EndOfCentralDirectoryRecord";
